Lab06/pyproject.toml
[project]
name = "lab07"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "matplotlib>=3.10.8",
    "numpy>=2.4.4",
    "pillow>=11.0.0",
]
